# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "نظام إدارة المكتبة الشخصية وتقييم الكتب (Personal Library API)",
        "description": "RESTful API متكامل لإدارة الكتب، المؤلفين، التصنيفات، وتقييمات الكتب مع توثيق كامل باستخدام Swagger.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "سيرفر التطوير المحلي (Local Development Server)"
        }
    ],
    "paths": {
...
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "استعراض قائمة الكتب مع الترقيم والبحث والفلترة",
                "operationId": "4770a9fbdee214124f78e5c87435016a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "البحث في اسم الكتاب أو الوصف",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "الفلترة حسب رقم التصنيف",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "author_id",
                        "in": "query",
                        "description": "الفلترة حسب رقم المؤلف",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "رقم الصفحة المراد استعراضها (الترقيم الصفحي)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "عدد العناصر في الصفحة الواحدة",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "قائمة الكتب المسترجعة بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "مقدمة ابن خلدون"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "كتاب يؤسس لعلم الاجتماع والتاريخ."
                                                    },
                                                    "cover_image": {
                                                        "type": "string",
                                                        "example": "storage/covers/abcdef.jpg"
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "author_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "user_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "تاريخ"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "author": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "ابن خلدون"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "إضافة كتاب جديد مع إمكانية رفع صورة الغلاف",
                "operationId": "1fe722c618c3c6300e645888c6cde4ec",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category_id",
                                    "author_id"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "عنوان الكتاب",
                                        "type": "string",
                                        "example": "مقدمة ابن خلدون"
                                    },
                                    "description": {
                                        "description": "وصف الكتاب",
                                        "type": "string",
                                        "example": "كتاب يؤسس لعلم الاجتماع والتاريخ"
                                    },
                                    "category_id": {
                                        "description": "معرّف التصنيف",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "author_id": {
                                        "description": "معرّف المؤلف",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "cover_image": {
                                        "description": "ملف صورة الغلاف (اختياري)",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "تمت إضافة الكتاب بنجاح",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "مقدمة ابن خلدون"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "كتاب يؤسس لعلم الاجتماع والتاريخ"
                                        },
                                        "cover_image": {
                                            "type": "string",
                                            "example": "storage/covers/abcdef.jpg"
                                        },
                                        "category_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "author_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "user_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "غير مصرح بالدخول (Bearer Token مفقود أو غير صالح)"
                    },
                    "422": {
                        "description": "خطأ في التحقق من البيانات (مثل تصنيف أو مؤلف غير موجود)"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "أدخل الـ Token الخاص بك هنا بالصيغة التالية: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}